modification nom des dossiers des controllers organisateurs et Participants
