+++ P2-BDE-Events/Controllers/OrganisateurControllers/EvenementAVenirController.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace P2_BDE_Events.Controllers.Organisateur
+namespace P2_BDE_Events.Controllers.OrganisateurControllers
 {
-    public class MessagesController : Controller
+    public class EvenementAVenirController : Controller
     {
         public IActionResult Index()
         {
