+++ P2-BDE-Events/Controllers/OrganisateurControllers/MesEvenementsController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace P2_BDE_Events.Controllers.Organisateur
+namespace P2_BDE_Events.Controllers.OrganisateurControllers
 {
     public class MesEvenementsController : Controller
     {
