+++ P2-BDE-Events/Controllers/OrganisateurControllers/MessagesController.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace P2_BDE_Events.Controllers.Organisateur
+namespace P2_BDE_Events.Controllers.OrganisateurControllers
 {
-    public class EvenementAVenirController : Controller
+    public class MessagesController : Controller
     {
         public IActionResult Index()
         {
