+++ P2-BDE-Events/Controllers/OrganisateurControllers/PrestatairesController.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace P2_BDE_Events.Controllers.Organisateur
+namespace P2_BDE_Events.Controllers.OrganisateurControllers
 {
-    public class MonBDEController : Controller
+    public class PrestatairesController : Controller
     {
         public IActionResult Index()
         {
