+++ P2-BDE-Events/Controllers/ParticipantControllers/MonCompteController.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace P2_BDE_Events.Controllers.Organisateur
+namespace P2_BDE_Events.Controllers.ParticipantControllers
 {
-    public class PrestatairesController : Controller
+    public class MonCompteController : Controller
     {
         public IActionResult Index()
         {
