+++ P2-BDE-Events/Models/Compte/Organisateur.cs
@@ -8,8 +8,6 @@ namespace P2_BDE_Events.Models.Compte
         public int Id { get; set; }
 
         public virtual Participant Participant { get; set; }
-
-
         public string FonctionBDE { get; set; }
 
     }
