+++ P2-BDE-Events/Views/Organisateur/EvenementsEnCours.cshtml
@@ -6,9 +6,9 @@
 
 @section Header {
     <ul>
-        <li onclick="window.location.href = '@Url.Action("EvenementsPrecedents", "MesEvenements", new { area = "Organisateur" })'">Evenements précédents</li>
-        <li class="pageActive" onclick="window.location.href = '@Url.Action("EvenementsEnCours", "MesEvenements", new { area = "Organisateur" })'">Evenements en cours</li>
-        <li onclick="window.location.href = '@Url.Action("EvenementsSuivants", "MesEvenements", new { area = "Organisateur" })'">Evenements suivants</li>
+        <li onclick="window.location.href = '@Url.Action("EvenementsPrecedents", "MesEvenements", new { area = "Organisateur" })'">Évènements précédents</li>
+        <li class="pageActive" onclick="window.location.href = '@Url.Action("EvenementsEnCours", "MesEvenements", new { area = "Organisateur" })'">Évènements en cours</li>
+        <li onclick="window.location.href = '@Url.Action("EvenementsSuivants", "MesEvenements", new { area = "Organisateur" })'">Évènements suivants</li>
     </ul>
 }
