+++ P2-BDE-Events/Views/Shared/_LayoutOrganisateur.cshtml
@@ -43,7 +43,7 @@
         <ul>
             <li><a onclick="window.location.href = '/Organisateur/MesEvenements/EvenementsEnCours'">Mes évènements</a></li>
             <li><a onclick="window.location.href = '/Organisateur/CreerUnEvenement/ChoixCreation'">Créer un évènement</a></li>
-            <li><a onclick="window.location.href = '/Organisateur/EvenementAVenir/'">Évènement à venir</a></li>
+            <li><a onclick="window.location.href = '/Organisateur/EvenementAVenir/'">Évènements à venir</a></li>
             <li><a onclick="window.location.href = '/Organisateur/Prestataires/'">Prestataires</a></li>
             <li><a onclick="window.location.href = '/Organisateur/Messages/'">Messages</a></li>
             <li><a onclick="window.location.href = '/Organisateur/MonBDE/'">Mon BDE</a></li>
